LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_multiple_pointerover_no_pointer_movement-expected.txt

initial position

FAIL Pointer boundary events should be fired per pointer assert_equals: expected "pointerover@div#target, pointerover@div#target, pointerover@div#container, pointerover@div#container" but got "pointerover@div#target"
